refactor(tools): improving folder structure
