+++ tools/src/docs/input/template.tmpl
